Introduce a ProductQueryService and move read operations from the database there
