+++ src/main/java/com/github/jbence1994/webshop/photo/ProductPhotoQueryServiceImpl.java
@@ -1,7 +1,7 @@
 package com.github.jbence1994.webshop.photo;
 
 import com.github.jbence1994.webshop.product.ProductNotFoundException;
-import com.github.jbence1994.webshop.product.ProductService;
+import com.github.jbence1994.webshop.product.ProductQueryService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -10,12 +10,12 @@ import java.util.List;
 @Service
 @AllArgsConstructor
 public class ProductPhotoQueryServiceImpl implements ProductPhotoQueryService {
-    private final ProductService productService;
+    private final ProductQueryService productQueryService;
     private final ProductPhotoRepository productPhotoRepository;
 
     @Override
     public List<ProductPhoto> getProductPhotos(Long productId) {
-        if (!productService.isProductExistById(productId)) {
+        if (!productQueryService.isProductExistById(productId)) {
             throw new ProductNotFoundException(productId);
         }
