+++ src/main/java/com/github/jbence1994/webshop/product/ProductQueryServiceImpl.java
@@ -1,0 +1,28 @@
+package com.github.jbence1994.webshop.product;
+
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+@AllArgsConstructor
+public class ProductQueryServiceImpl implements ProductQueryService {
+    private final ProductRepository productRepository;
+
+    @Override
+    public List<Product> getProducts() {
+        return productRepository.findAll();
+    }
+
+    @Override
+    public Product getProduct(Long id) {
+        return productRepository.findById(id)
+                .orElseThrow(() -> new ProductNotFoundException(id));
+    }
+
+    @Override
+    public boolean isProductExistById(Long id) {
+        return productRepository.existsById(id);
+    }
+}
