+++ src/main/java/com/github/jbence1994/webshop/product/ProductServiceImpl.java
@@ -3,28 +3,10 @@ package com.github.jbence1994.webshop.product;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
 @Service
 @AllArgsConstructor
 public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
-
-    @Override
-    public List<Product> getProducts() {
-        return productRepository.findAll();
-    }
-
-    @Override
-    public Product getProduct(Long id) {
-        return productRepository.findById(id)
-                .orElseThrow(() -> new ProductNotFoundException(id));
-    }
-
-    @Override
-    public boolean isProductExistById(Long id) {
-        return productRepository.existsById(id);
-    }
 
     @Override
     public Product createProduct(Product product) {
